refactor: Rename title + main menu
Home screen now called main menu
Previous main menu now called title screen
