+++ CMakeLists.txt
@@ -1120,15 +1120,15 @@ add_executable(group2
         src/client/menus/lobby/Lobby.cpp
         src/client/menus/lobby/ui/LobbyUI.hpp
         src/client/menus/lobby/ui/LobbyUI.cpp
+        src/client/menus/title/TitleScreen.hpp
+        src/client/menus/title/TitleScreen.cpp
+        src/client/menus/title/ui/TitleScreenUI.hpp
+        src/client/menus/title/ui/TitleScreenUI.cpp
+        src/client/renderer-new/RendererTypes.hpp
         src/client/menus/main/MainMenu.hpp
         src/client/menus/main/MainMenu.cpp
         src/client/menus/main/ui/MainMenuUI.hpp
         src/client/menus/main/ui/MainMenuUI.cpp
-        src/client/renderer-new/RendererTypes.hpp
-        src/client/menus/home/Home.hpp
-        src/client/menus/home/Home.cpp
-        src/client/menus/home/ui/HomeUI.hpp
-        src/client/menus/home/ui/HomeUI.cpp
         src/client/menus/MenuTheme.hpp
         src/client/menus/MenuTheme.cpp
         src/ecs/components/PowerupSpawner.hpp
