+++ src/client/menus/main/ui/MainMenuUI.hpp
@@ -1,18 +1,49 @@
 /// @file MainMenuUI.hpp
-/// @brief ImGui widget for the top-level landing menu.
+/// @brief ImGui widget for the main menu server join form.
 
 #pragma once
 
-/// @brief Output from a single main-menu UI frame.
-struct MainMenuResult
+#include "network/DiscoveryClient.hpp"
+#include "network/discovery/GlobalDiscoveryProtocol.hpp"
+
+#include <string>
+#include <string_view>
+#include <vector>
+
+/// @brief Mutable widget state for the server join form.
+struct JoinMenuState
 {
-    bool playClicked = false; ///< True if the user pressed "Play" this frame.
-    bool hostClicked = false; ///< True if the user pressed "Host" this frame.
-    bool exitClicked = false; ///< True if the user pressed "Exit" this frame.
+    std::string serverIp = "127.0.0.1"; ///< Server hostname or IP address entered by the user.
+    int serverPort = 2310;              ///< Server port entered by the user.
+};
+
+/// @brief Output from a single main menu UI frame.
+struct JoinMenuResult
+{
+    bool connectClicked = false;             ///< True if the user pressed "Join" this frame.
+    bool hostClicked = false;                ///< True if the user pressed "Host" this frame.
+    bool refreshClicked = false;             ///< True if the user requested a global browser refresh.
+    bool returnToTitleScreenClicked = false; ///< True if the user pressed "Return to Title Screen" this frame.
+    int localServerIndex = -1;               ///< Index of a discovered local server to join, or -1.
+    int globalServerIndex = -1;              ///< Index of a discovered server to join, or -1.
 };
 
 namespace main_menu_ui
 {
-/// @brief Render the landing menu and return any user action this frame.
-MainMenuResult buildMainMenu();
+
+/// @brief Render the join game window and return any user action this frame.
+/// @param state Persistent widget state (IP/port fields).
+/// @param errorMessage Optional error string displayed in red below the form.
+/// @param localServers LAN-discovered servers shown in the local browser table.
+/// @param globalServers Directory-discovered servers shown in the global browser table.
+/// @param browserError Optional global-browser error displayed beside the refresh button.
+/// @param browserRefreshing True while a global-browser refresh is in flight.
+/// @return Actions the caller should apply (join, host, or refresh request).
+JoinMenuResult buildJoinMenu(JoinMenuState& state,
+                             std::string_view errorMessage,
+                             const std::vector<DiscoveryClient::DiscoveredServer>& localServers,
+                             const std::vector<net::discovery::ServerInfo>& globalServers,
+                             std::string_view browserError,
+                             bool browserRefreshing);
+
 } // namespace main_menu_ui
